Introduce mechanism for loading content from YAML files
This commit introduces the load_content function that will retrieve
structured content from a YAML file and make it available as a Python
object, crucially with properties available via dot notation.

The content obtained by load_content is then passed to the template
via render_template.

The intent of using YAML to store the content is that:

1. It allows us to keep all the content together and structured in
   a way that reflects its inherent organisation
2. It avoids the need for content to be stored in Python code or
   HTML templates, making content review and management easier.