Add HTML templates for the form and submitted pages
The request-a-service-record.html template makes use of content
stored in YAML and passes params to the form elements to affect
their presentation (specifically, changing the heading level to
something that will work with the form and setting the size of
text inputs).

The app initialisation is also updated to include WTFormsHelpers
because these are needed to translate the errors object on the
form passed to the view (form.errors) into a structure that is
expected by tnaErrorSummary.

Because submitted.html contains content for testing/debugging
purposes, this has not been included in the YAML content, and is
instead included directly in the template.