Remove fieldsets from the request form
An earlier commit introduced fieldsets to group related fields but
I've since realised - thanks to a comment in the PR[^1] - that some
fields output by our forms library come pre-wrapped in fieldsets (
an example would be the date inputs).

Nesting of fieldsets is known to cause issues for some users of
assistive technology[^2]

  "It is possible to nest one <fieldset> element inside another,
   but it is not recommended. Screen readers do not automatically
   indicate the end of the <fieldset> element, so it is impossible
   for screen reader users to confidently know which fields belong
   within which fieldset."

Because of this, I feel it is better to remove the fieldsets I'd
introduced and instead rely on the overall page structure.

[^1]: https://github.com/nationalarchives/ds-request-service-record/pull/1#discussion_r2219128393
[^2]: https://accessibility.blog.gov.uk/2016/07/22/using-the-fieldset-and-legend-elements/