Add and update fields
This commit adds two new fields (country and postcode) using the
autocomplete tokens from the WHATWG spec[^1].

For the time-being all form options are included in content.yaml,
so I've done the same here for country. My sense is that a good
future refactor will be to move these out of the content - but
it works for now.

I've tried to make the code a bit easier to read by introducing
a prepare_country_options helper. I haven't applied this helper
to other fields because countries appear to be unique in that
the value is the same as the text presented to the user.

[^1]: https://html.spec.whatwg.org/multipage/form-control-infrastructure.html#autofill-expectation-mantle