Add a basic RoutingStateMachine
This basic machine has:

- five states. I'm using the `value` property of State instances to store
  the next route
- four events

I'm still very much finding my way with state machines but I've opted to have
all the states progressed to from 'initial'. I've done this because HTTP is
stateless so, at the point of a request being received, the app will be in an
'initial' state. It may be that future states can be progress to within a route
function (I can imagine this being the case for the `main.send_to_gov_pay` route,
for example), but I want to keep this super simple for the time being.

The