Refactor start route to separate GET and POST methods
For some reason - which I don't fully understand, if I'm honest - having both POST
and GET handled by the same route function was causing the Playwright tests to return
a 200 when the user submitted the form (rather than the redirect to the next page).

I spent a good few hours today trying to get to the bottom of this and eventually found
the only thing which made Playwright get the same result as the manual browser journey
(in a fresh browser, with no cookies) was to split the journey in this way.